content/media/test/test_closing_connections.html

<!DOCTYPE HTML>
<html>



<head>
  <title>Test for Bug 479863 --- loading many connections</title>
  <script type="application/javascript" src="/MochiKit/packed.js"></script>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
  <script type="application/javascript" src="use_large_cache.js"></script>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=479863">Mozilla Bug 479863</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>

<script type="application/javascript">
window.onload = function() {
  ok(true, "loaded metadata for all videos");
  SimpleTest.finish();
}
</script>








<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>
<video src="seek.ogv"></video>

<pre id="test">
<script type="application/javascript">
SimpleTest.waitForExplicitFinish();
</script>
</pre>
</body>
</html>
